Sector loading now works, I think

+++ server/main.c
@@ -38,7 +38,7 @@ int main(){
 		}
 		sendInfo();
 		clock_gettime(CLOCK_MONOTONIC, &otherTime);
-		long int sleep = (long int)25000000 - (otherTime.tv_nsec-lastTime.tv_nsec) - 1000000000l*(otherTime.tv_sec-lastTime.tv_sec);
+		int32_t sleep = (int32_t)25000000 - (otherTime.tv_nsec-lastTime.tv_nsec) - 1000000000l*(otherTime.tv_sec-lastTime.tv_sec);
 		if(sleep > 0){
 			t.tv_nsec = sleep;
 			nanosleep(&t, NULL);
