+++ server/searchforsector.c
@@ -1,6 +1,6 @@
 #include "globals.h"
 #include <stdlib.h>
-sector *searchforsector(long long int x, long long int y){
+sector *searchforsector(uint64_t x, uint64_t y){
 	sector *conductor;
 	conductor = listrootsector;
  	while(conductor != NULL && (conductor->x != x || conductor->y != y)){
