+++ server/unloadsector.c
@@ -5,12 +5,12 @@
 #include "globals.h"
 int unloadsector(sector *target){
 	writesectortofile(target->x, target->y);
-	printf("unloadsector called %lld, %lld\n", target->x, target->y);
+	printf("unloadsector called %ld, %ld\n", target->x, target->y);
 	//free all entities
 	free(target);
 	return(0);
 }
-int writesectortofile(long long int x, long long int y){
+int writesectortofile(uint64_t x,uint64_t y){
 	//writedatshit
 	return(0);
 }
